Change dropdown component
