+++ bigbluebutton-html5/imports/ui/components/dropdown/styles.scss
@@ -1,67 +1,7 @@
 @import "../../stylesheets/variables/_all";
 
-.settingsMenuItem {
-  padding-top:20px;
-}
+.dropdown {}
 
-.settingsMenuItemText {
-  margin-left: 10px;
-}
+.content {}
 
-.menuList {
-  list-style-type: none;
-  padding-left: 0px;
-}
-
-.settingBtn {
-  -ms-transform: rotate( 90deg ); /* IE 9 */
-  -webkit-transform: rotate( 90deg ); /* Safari */
-  transform: rotate( 90deg );
-  color: #ffffff;
-  span {
-    border: 0px solid;
-    box-shadow: none;
-  }
-}
-
-.hrDropdown {
-  border: 0.3px solid $color-gray-light;
-  width: 165px;
-  text-align: center;
-}
-
-.dropdown {
-    position: relative;
-}
-
-.dropdownActiveContent {
-  margin-top: 14px;
-  margin-right: -3px;
-  padding-left: 15px;
-  height: auto;
-  width: 190px;
-  display: block;
-  background-color: #ffffff;
-  position: absolute;
-  right: 0;
-  text-align: left;
-  font-size: $font-size-base;
-  border-radius: 3%;
-  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
-}
-
- .triangleOnDropdown {
-   position: absolute;
-   margin-top: 5px;
-   margin-left: -2px;
-   width: 0;
-   height: 0;
-   border-left: 18px solid transparent;
-   border-right: 18px solid transparent;
-   border-bottom: 18px solid #ffffff;
- }
-
- .iconColor {
-   font-size: $font-size-base;
-   color: $color-gray-light;
- }
+.trigger {}
